master/requirements.txt
affine==2.3.0
attrs==21.2.0
brotlipy==0.7.0
certifi==2020.12.5
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
GDAL==3.0.0
grpcio==1.38.1
grpcio-tools==1.38.1
numpy==1.21.0
Pillow==8.3.1
protobuf==3.17.3
rasterio==1.2.6
